Revert logic separation of map capture server.
